Added the price display

+++ db.js
@@ -21,7 +21,7 @@ function initializeDb() {
       id INTEGER PRIMARY KEY AUTOINCREMENT,
       name TEXT NOT NULL UNIQUE,
       category TEXT DEFAULT '',
-      price INTEGER NOT NULL DEFAULT 0,
+      price FLOAT NOT NULL DEFAULT 0,
       created_at DATETIME DEFAULT CURRENT_TIMESTAMP
     );
